Provide a mechanism to test api calls.
Expose PivotalTracker::API::TestClient as a replacement of normal client. It
allows to use vcr and webmock to avoid hitting API multiple times.

As a requirement, VCR needs to be setup before this can run properly.
