+++ spec/spec_helper.rb
@@ -2,10 +2,9 @@ require "codeclimate-test-reporter"
 CodeClimate::TestReporter.start
 
 require 'pivotaltracker'
-require 'vcr'
 require 'byebug'
-require 'webmock'
 require 'support/test_api_setup'
+require 'pivotaltracker/api/test_client'
 
 RSpec.configure do |config|
   config.expect_with :rspec do |expectations|
@@ -21,25 +20,4 @@ VCR.configure do |config|
   config.cassette_library_dir = File.join(__dir__, 'fixtures', 'vcr_cassettes')
   config.hook_into :webmock
   config.ignore_hosts 'codeclimate.com'
-end
-
-module PivotalTracker
-  class API
-    class TestClient < ::PivotalTracker::API::Client
-      def get(endpoint, id, options={})
-        casette =  if id < 0
-          "non_existing"
-        else
-          "#{id}"
-        end
-
-        resource = endpoint.gsub('/', '')
-        casette.prepend("#{resource}")
-
-        VCR.use_cassette(casette) do
-          super(endpoint, id, options)
-        end
-      end
-    end
-  end
 end
